feat: replace ic-mgmt by bitcoin canister to get Utxos
