+++ package.json
@@ -35,7 +35,7 @@
 		"@dfinity/agent": "^1.3.0",
 		"@dfinity/auth-client": "^1.3.0",
 		"@dfinity/candid": "^1.3.0",
-		"@dfinity/ckbtc": "^2.3.3",
+		"@dfinity/ckbtc": "^2.3.3-next-2024-06-04",
 		"@dfinity/cketh": "^3.0.1-next-2024-05-21.2",
 		"@dfinity/gix-components": "^4.2.0-next-2024-05-06.2",
 		"@dfinity/ic-management": "^4.0.0",
